Student Application Project
I had a chance to create classes initially and use them modules, by creating new files with the names school.js and student.js. In another file named database.js, I practiced the fs module. For a full implementation of this project, chalk needs to be installed, though it has nothing to do with the project and I only used it for practice. 

This project has been created as part of the homework for Women Techmakers Berlin JS Crash Course published on Armagan Amcalar's YouTube channel.